+++ lib/src/app.dart
@@ -159,7 +159,8 @@ class App<T extends Context> {
           if (error is HttpError) {
             _responseHttpError(ctx, error);
           } else {
-            final err = HttpError(500, stackTrace: stackTrace, error: error);
+            final err =
+                HttpError(500, stackTrace: stackTrace, exception: error);
             _responseHttpError(ctx, err);
           }
         });